Terrain and Textures Update
to get things moving I've deactivated the Ground and Ground_Mesh from the demo playground under Environment > Greybox > Primitives and added in a terrain. I've also added a few textures and made them into materials so we have more to work with in terms of our theme. I've also gotten started on a scene called Village Pre Alpha but didn't get anywhere other than slapping the terrain down, the player and a camera.